datasets/cloud_data_connector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud_data_connector"
version = "1.0.2"
requires-python = ">=3.9,<3.11"
authors = [
    { name="IntelAI", email="IntelAI@intel.com"}
]
description = "Intel's cloud data connector"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent"
]
dependencies = [
    "azureml>=0.2.7",
    "azure-ai-ml>=1.4.0",
    "azureml-core>=1.49.0",
    "azure-identity>=1.12.0",
    "azure-storage-blob>=1.4.1",
    "azure-core>=1.26.3",
    "boto3>=1.26.154",
    "google-api-core>=2.0.0",
    "google-auth>=1.33.0",
    "google-auth-oauthlib>=0.4.1",
    "google-cloud-bigquery>=2.1.0",
    "google-cloud-storage>=2.1.0",
    "packaging>=21.3",
    "python-dotenv>=1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["cloud_data_connector*"]
exclude = ["cloud_data_connector.egg-info"]
namespaces = false
